bridge/requirements.txt
hangups
pyotp
discord.py
lockfile
